Move devServer to dev-server
