+++ packages/webpack-common/src/index.js
@@ -1,4 +1,3 @@
-export { default as devServer } from './devServer'
 export { default as parseVersion } from './parseVersion'
 export * from './extractText'
 export * from './webpack'
